requirements.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.39.1
django-filter==2.0.0
django-rest-auth==0.9.3
django-sendsms==0.3.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
drfaddons==0.1.3
idna==2.8
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pillow==5.4.1
pycrypto==2.6.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.24.1
